IDP-1004 @Duration can go on a method.  Add to appropriate methods
https://issues.shibboleth.net/jira/browse/IDP-1004

See also https://jira.spring.io/browse/SPR-14434

In order to make the annotation-driven converted work robustly we
need to add @Duration to the setter for any @Duration the appropriate
parameter in the constructors.

Make @Duration usable in such instances.

Add @Duration to all appropriate setters as well as their associated
getters.  Leave the @Duration on the parameter for readability.