JSPT-75 Do not text inequivalent things
https://issues.shibboleth.net/jira/browse/JSPT-75

In the pretty print test compare the pretty printed output with other
pretty printed output.

Leave the only one (straight extrusion of the XML as text against pretty
print output in place for pre V9 versions.  Test for V9 by the fact that
its the first version of java to _BNOT_) be major version 1.  Ugly, but
functional.